requirements.txt
tensorflow>=2.15.0
keras>=2.15.0
streamlit>=1.31.0
numpy>=1.24.0
Pillow>=10.0.0
tensorflow-hub>=0.15.0
bing_image_downloader>=1.1.2 